% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/givitiCalibrationMain.R
\name{givitiCalibrationBeltTable}
\alias{givitiCalibrationBeltTable}
\title{Table of the Calibration Belt Significant Deviations}
\usage{
givitiCalibrationBeltTable(cb, tableStrings, grayLevels, xlim, ylim)
}
\arguments{
\item{cb}{A \code{givitiCalibrationBelt} object, to be generated with
the function \code{givitiCalibrationBelt}.}

\item{tableStrings}{Optional. A list with four character elements named
\code{overBisString},\code{underBisString},\code{confLevelString},
\code{neverString}. The four strings of the list are printed instead of the
texts "Over the bisector"/"Under the bisector"/"Confidence level"/"NEVER"
in the table reporting the intersections of the calibration belt with
the bisector.}

\item{grayLevels}{A vector containing the code of the
gray levels used in the plot of the calibration belt.}

\item{xlim,ylim}{Numeric vectors of length 2, giving the
x and y coordinates ranges. Default values are \code{c(0,1)}.}
}
\value{
The function prints the table on the graphical area.
}
\description{
\code{givitiCalibrationBeltTable} prints on the graphical area of the calibration
 belt plot the table that summarizes the significant deviations from the
 line of perfect calibration (i.e. the bisector of the I quadrant).
}

