% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/givitiCalibrationMain.R
\name{givitiCheckArgs}
\alias{givitiCheckArgs}
\title{Check of the argument's values}
\usage{
givitiCheckArgs(o, e, devel, thres, maxDeg)
}
\arguments{
\item{o}{A numeric vector representing the binary outcomes.
The elements must assume only the values 0 or 1. The predictions
in \code{e} must represent the probability of the event
coded as 1.}

\item{e}{A numeric vector containing the probabilities of the
model under evaluation. The elements must be numeric and between 0 and 1.
The lenght of the vector must be equal to the length of the vector \code{o}.}

\item{devel}{A character string specifying if the model has been fit on
the same dataset under evaluation (\code{internal}) or if the model has
been developed on an external sample (\code{external}).}

\item{thres}{A numeric scalar between 0 and 1 representing 1 - the significance level
adopted in the forward selection.}

\item{maxDeg}{The maximum degree considered in the forward selection.}
}
\value{
The function produce an error if the elements provided
 through the arguments do not meet the constraints reported.
}
\description{
Check of the coherence of the values passed to the functions
\code{givitiCalibrationTest} and \code{givitiCalibrationBelt}.
}

