% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/givitiCalibrationBasic.R
\name{logit}
\alias{logistic}
\alias{logit}
\title{Logit and logistic functions}
\usage{
logit(p)

logistic(x)
}
\arguments{
\item{p}{A numeric vector whose components are numbers between 0 and 1.}

\item{x}{A numeric vector.}
}
\value{
The functions apply the logit and logistic transformation to each element of the vector
passed as argument.
In particular, logit(p)=ln(p/(1-p)) and logistic(x)=exp(x)/(1+exp(x)).
}
\description{
\code{logit} and \code{logistic} implement the logit and logistic transformations, respectively.
}
\examples{
logit(0.1)
logit(0.5)
logistic(0)
logistic(logit(0.25))
logit(logistic(2))
}

